+++ src/Calculators/Dominion/SpellCalculator.php
@@ -64,7 +64,7 @@ class SpellCalculator
 
         // Heroes
         if ($this->spellHelper->isInfoOpSpell($spell)) {
-            if ($dominion->hero !== null && $dominion->hero->getPerkValue('info_spell_cost')) {
+            if ($dominion->hero !== null) {
                 $spellCostMultiplier += $dominion->hero->getPerkMultiplier('info_spell_cost');
             }
         }